# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Book API",
    "description": "API for managing books",
    "version": "1.0.0"
  },
  "paths": {
    "/books/": {
      "get": {
        "tags": ["books"],
        "summary": "Read Books",
        "operationId": "read_books_books__get",
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "default": 0, "title": "Skip" }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "default": 10, "title": "Limit" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/BookRead" },
                  "title": "Response Read Books Books  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["books"],
        "summary": "Create Book",
        "operationId": "create_book_books__post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/BookCreate" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/BookRead" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/books/{book_id}": {
      "get": {
        "tags": ["books"],
        "summary": "Read Book",
        "operationId": "read_book_books__book_id__get",
        "parameters": [
          {
            "name": "book_id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "title": "Book Id" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/BookRead" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["books"],
        "summary": "Update Book",
        "operationId": "update_book_books__book_id__put",
        "parameters": [
          {
            "name": "book_id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "title": "Book Id" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/BookUpdate" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/BookRead" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["books"],
        "summary": "Delete Book",
        "operationId": "delete_book_books__book_id__delete",
        "parameters": [
          {
            "name": "book_id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "title": "Book Id" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/BookDelete" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/": {
      "get": {
        "summary": "Root",
        "operationId": "root__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "BookCreate": {
        "properties": {
          "title": { "type": "string", "title": "Title" },
          "author": { "type": "string", "title": "Author" },
          "description": { "type": "string", "title": "Description" }
        },
        "type": "object",
        "required": ["title", "author", "description"],
        "title": "BookCreate"
      },
      "BookDelete": {
        "properties": { "id": { "type": "integer", "title": "Id" } },
        "type": "object",
        "required": ["id"],
        "title": "BookDelete"
      },
      "BookRead": {
        "properties": {
          "title": { "type": "string", "title": "Title" },
          "author": { "type": "string", "title": "Author" },
          "description": { "type": "string", "title": "Description" },
          "id": { "type": "integer", "title": "Id" }
        },
        "type": "object",
        "required": ["title", "author", "description", "id"],
        "title": "BookRead"
      },
      "BookUpdate": {
        "properties": {
          "title": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Title"
          },
          "author": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Author"
          },
          "description": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Description"
          }
        },
        "type": "object",
        "title": "BookUpdate"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": { "$ref": "#/components/schemas/ValidationError" },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": { "anyOf": [{ "type": "string" }, { "type": "integer" }] },
            "type": "array",
            "title": "Location"
          },
          "msg": { "type": "string", "title": "Message" },
          "type": { "type": "string", "title": "Error Type" }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      }
    }
  }
}